+++ fast_barcode_scanner_platform_interface/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Camera position now taken into account on Android.
+* Fix DetectionMode.pauseDetection on Android.
+* Updated CameraX to v1.1.0-alpha05
+
 ## 1.0.2
 
 * Even more documentation.